requirements.txt
Flask
Flask-Bootstrap
Flask-WTF
Flask-SQLAlchemy
requests
gunicorn